move more commands
