+++ src/commands/mod.rs
@@ -1,7 +1,6 @@
-// pub mod join;
 pub mod leave;
 pub mod pause;
 pub mod ping;
+pub mod skip;
 pub mod stop;
-// pub mod play;
-// pub mod skip;
+pub mod unpause;
