+++ src/commands/ping.rs
@@ -2,7 +2,9 @@ use crate::{check_msg, CommandResult, Context};
 
 #[poise::command(slash_command, guild_only)]
 pub async fn ping(ctx: Context<'_>) -> CommandResult {
-    let reply = poise::CreateReply::default().content("Pong!");
+    let reply = poise::CreateReply::default()
+        .content("Pong!")
+        .ephemeral(true);
     check_msg(ctx.send(reply).await);
     Ok(())
 }
